mark key_hash/key_map/light_array to internal and some simple refactor
