+++ cpp/src/arrow/compute/key_map_internal.h
