+++ cpp/src/arrow/compute/key_map_internal_avx2.cc
@@ -17,7 +17,7 @@
 
 #include <immintrin.h>
 
-#include "arrow/compute/key_map.h"
+#include "arrow/compute/key_map_internal.h"
 #include "arrow/util/logging.h"
 
 namespace arrow {
