+++ cpp/src/arrow/compute/light_array_internal.h
